+++ platforms/nixos/modules/nmasur/presets/services/immich/immich.nix
@@ -10,6 +10,18 @@ in
   options.nmasur.presets.services.immich.enable = lib.mkEnableOption "Immich photo manager";
   config = lib.mkIf cfg.enable {
 
+    secrets.immich-oidc-secret = {
+      source = ./immich-oidc-secret.age;
+      dest = "${config.secretsDirectory}/immich-oidc-secret";
+      owner = config.services.immich.user;
+      group = config.services.immich.group;
+      permissions = "0440";
+    };
+    systemd.services.immich-oidc-secret-secret = {
+      requiredBy = [ "immich-server.service" ];
+      before = [ "immich-server.service" ];
+    };
+
     services.immich = {
       enable = true;
       port = 2283;
@@ -19,13 +31,28 @@ in
       machine-learning.environment = { };
       mediaLocation = "/data/images";
       secretsFile = null;
-      settings.server.externalDomain = "https://${hostnames.photos}";
+      settings = {
+        server.externalDomain = "https://${hostnames.photos}";
+        oauth = {
+          enabled = true;
+          issuerUrl = "https://${hostnames.auth}";
+          clientId = "1f4e0f8d-6cee-4d67-8d53-74bf6c18ae09";
+          clientSecret._secret = config.secrets.immich-oidc-secret.dest;
+          scope = "openid profile email";
+          autoRegister = true;
+          buttonText = "Login with Pocket ID";
+        };
+      };
       environment = {
         IMMICH_ENV = "production";
         IMMICH_LOG_LEVEL = "log";
         NO_COLOR = "false";
         IMMICH_TRUSTED_PROXIES = "127.0.0.1";
       };
+    };
+
+    systemd.services.immich-server = {
+      after = [ "immich-oidc-secret-secret.service" ];
     };
 
     nmasur.presets.services.caddy.routes = [